edit more saving score

+++ ButtonTap/HighScoreViewController.h
@@ -10,6 +10,6 @@
 
 @interface HighScoreViewController : UITableViewController
 {
-    NSInteger highscore;
+    int highscore;
 }
 @end
